Add the 'as' field to ImageSource to allow multi-stage builds
The `as` field allows image sources to be used directly by multi-stage
Docker builds to replace hardcoded strings in the Dockerfile with images
delivered via image streams. After this change, both `paths` and `as`
are optional as long as at least one is specified. The list of values in
the `as` field are exact string replacements for fields in the
Dockerfile.