Add IdentityMetadata API
This API is used during flows that require a cookie based session to
store IDP state.  Since cookies are limited in size, it is required
to make sure that all groups can be accounted for.  The name of the
provider is stored to allow us to qualify groups based on the
provider.  This prevents any conflicts between identity providers.
Since login flows can fail, this API uses an expiration field that
is backed by an etcd lease.  This ensures that these objects are
cleaned up (in the same manner as OAuth tokens).

The OAuth token APIs were updated with matching fields (the name of
the provider and the groups).  This ties each login to the set of
groups asserted by the provider at that time.

Signed-off-by: Monis Khan <mkhan@redhat.com>