API changes for groups from IDPs
This change adds the unqualifiedGroups field to the identity
provider struct.  This field is used to determine if groups from
IDPs should be used as-is or if they should be qualified using the
prefix openshift:idp:groups:<idp_name>: (this prefix prevents any
conflicts between IDPs).

As a starting point, the request header and OIDC IDPs were updated
with the relevant fields required to assert group memberships.

Signed-off-by: Monis Khan <mkhan@redhat.com>