Update ClusterVersion with a desired field and a history field
The current field is not useful because it doesn't represent level
driven flows well. Instead, track the version the cluster is working
towards as `status.desired` and each time that target changes update
the `status.history` field. The `status.history` field allows us to
know what previous updates were made and what times those completed.

Even though rollback is not desirable, we need to report a field that
indicates what the last successfully applied version was (user starts
at 4.0.0, then upgrades to 4.0.1, which fails, then a new 4.0.2
version is released that also fails, we need to report 4.0.0 as the
version to go back to).