clusterversion: Rename 'Payload' to 'Image'
We consistently use the "release image" terminology in our commands
and design docs when we describe the artifact that carries our manifests
in an update. Payload predates that term (from Omaha) and general
consensus is that we have no plans to deliver our updates in any form
other than as an OCI compatible image.

Update the cluster version object to make it clear that an API consumer
must interpret the payload as a standard Docker image, as they already
do.