Remove useUsernameIdentity from KeystoneIdentityProvider
This field is only necessary if we plan on supporting 3.11 to 4.0
upgrades.  Since that is currently not the case, it should be
removed from the top level config.  Note that this does not remove
any of the wiring in osin that is required to support the legacy
path.  Thus if we ever decide to support 3.11 to 4.0 upgrades, we
will add this field back and rewire it through the operator.

Signed-off-by: Monis Khan <mkhan@redhat.com>