config/v1: add public and private zones to DNSSpec
Based on [1] `DNS` should be the source of truth for operators that manage DNS for their components.
`DNS` currently specifies the `BaseDomain` that should be used to make aure all the records after subdomains to `BaseDomain`.

A missing piece for operators that need to create DNS records is where should these records be created. For example, the `cluster-ingress-operator`
creates DNS records in public and private r53 zones on AWS by listing all zones that match `BaseDomain` [2]. The ingress operator is currently making an
assumption that the public zone matching the `BaseDomain` is *probably* the correct zone.

With changes in installer [3] of creating private r53 zone `cluster_name.base_domain` and using public zone `base_domain`. The `BaseDomain` in `DNSSpec` will
be set to the `cluster_domain` (`cluster_name.base_domain`) as all records must be subdomain of the `cluster_domain`. This breaks the previous assumption for the `cluster-ingress-operator`
or any other operator.

Clearly there is a gap to be filled regarding where the DNS records should be created. The installer knows which public and private zones should be used and can provide that information for the operators.

`DNSSpec` is extended to include a required `PrivateZone` field and an optional `PublicZone` field to provide operators location of where the corresponding records should be created.

`DNSZone` struct is also added to allow defining the DNS zone either using an `ID` or a string to string map `Tags`. `ID` allows installer to specify the public zone as it is predetermined, while `Tags` allow
installer to point to a private DNS zone that will be created for the cluster.

[1]: https://github.com/openshift/api/blob/d67473e7f1907b74d1f27706260eecf0bc9f2a52/config/v1/types_dns.go#L9
[2]: https://github.com/openshift/cluster-ingress-operator/blob/e7517023201c485428b3cdb3a86612230cf49e0a/pkg/dns/aws/dns.go#L111
[3]: https://github.com/openshift/installer/pull/1169