Add route.spec.subdomain to support implicit hostnames
Setting `spec.host` requires a user know the exact subdomain of
the ingress controller that will be used which is not portable
across clusters. For infrastructure components we want to have
consistent default names like `openshift-console`,
`openshift-registry`, etc which shouldn't require the operator to
introspect the default ingress controller or global config.

Instead, add a new optional field that if host is unset is
interpreted as a suggestion to the ingress controller to place
a name within its space. This field, like host, may only be
changed post creation by a user with permission since it has
the same ordering problems across namespaces. `subdomain`
can be set to `openshift-console` and the router should add
`apps.foo.com` to it (instead of following the default rule
`NAME-NAMESPACE`). Since it is a new field, its use is
optional and routers may ignore it in practice.