config: Clarify operand version for release payload
Fix a false statement in the godoc for OperandVersion.

Before this change, the godoc said that version must match the Available
condition, which is a false statement in the case of a manifest in a
release payload, which necessarily does *not* have an Available condition but
*must* have version set.  Following the old godoc literally will trigger
fallback mode in CVO.

This commit changes the godoc to state (1) that version must match
Available only in the case that an operator sets version and (2) that
version is allowed not to match Available only in the case that it is in a
release payload manifest.

* config/v1/types_cluster_operator.go (OperandVersion): Clarify that
version need not match the Available condition in a release payload
manifest.
* config/v1/zz_generated.swagger_doc_generated.go: Regenerate.