config/v1/types_infrastructure: Split apiServerInternalURL from apiServerURL
The kubelet (via the machine-config operator [1]) and Kubernetes
API-server operator [2] need the internal API URI.  The console
operator [3] and authentication operator [4] need the external URI.
Since the api-int split (e.g. openshift/installer@13e4b702f7,
data/aws: create an api-int dns name, 2019-04-11,
openshift/installer#1601), those are two different things, so we need
two different properties.

I'd originally proposed deprecating the old property because it was
ambiguous, but Clayton suggested using the generic name for the
external URI because the internal URI is only intended for a small
subset of components [5].

I've avoided leading with Internal so these properties will all
alphabetize together when rendering YAML, etc.

I've avoided "Private", because that is often used in a public/private
dichotomy for security.  The split was motivated by separating X.509
chains of trust, but the URIs themselves are not sensitive.

I'd originally proposed changing URL to URI, because we don't need to
make a point of locator-ness and RFCs recommend using the more generic
label [5,6].  But Sam and Clayton prefer sticking with URL, because it
is used more widely within OpenShift [7,8].

[1]: https://github.com/openshift/machine-config-operator/blob/2bd29e5005fb6f438f1b1c512176c3f7e5cf9e47/pkg/operator/operator.go#L393
[2]: https://github.com/openshift/cluster-kube-apiserver-operator/blob/cbcb3f403f674dd8d0b590d770c0ebf6ecb75f17/pkg/operator/certrotationcontroller/externalloadbalancer.go#L18
[3]: https://github.com/openshift/console-operator/blob/fc721a6990b6ca27efb60062cf5e9d620ec0ab01/pkg/console/subresource/configmap/configmap.go#L59
[4]: https://github.com/openshift/cluster-authentication-operator/blob/96643e4b29452f9ad20f7b90fc0794ffc3006fa2/pkg/operator2/oauth.go#L117
[5]: https://github.com/openshift/api/pull/308#discussion_r281268653
[5]: https://tools.ietf.org/html/rfc3305#section-2.2
[6]: https://tools.ietf.org/html/rfc3986#section-1.1.3
[7]: https://github.com/openshift/api/pull/308#discussion_r281263673
[8]: https://github.com/openshift/api/pull/308#discussion_r281263673