config/v1/types_infrastructure: Add BareMetal status for BareMetal platform
This adds a BareMetal-specific status fields, akin to AWS status.
It holds the API IP, the DNS IP, and the Ingress IP.

The API IP is provided by the installer. It is a reserved IP that the
installer provides that will first reside on the bootstrap node, and move
to one of the masters once the cluster is functioning.

The DNS IP will front a DNS service running on the bootstrap, and then the
cluster itself, that will serve DNS internal to the cluster, and separate
from the main cluster DNS service. It will serve etcd SRV records, as well
as hostnames for the BareMetal hosts.

The Ingress IP will be held by a node that runs an OpenShift router pod. An
external datacenter DNS record will point the apps wildcard record there.