config/v1: Introduce operatorhub API
Add the operatorhub.config.openshift.io API that can be used to enable
and disable default OperatorSources on the cluster. Disabling an
OperatorSource will remove the object from the cluster and enabling an
OperatorSource will add the object to the cluster using the default
manifest that is part of the operator image.

Ran "make generate-with-container"