Add KuryrConfig.OpenStackServiceNetwork for CNO
Trying to apply OpenStack Octavia to OpenShift networking models we
realized that the subnet for OpenShift Services created in OpenStack
needs to be twice the size of the subnet used by OpenShift, as Octavia
uses 2 IP's per loadbalancer - VIP (i.e. the one assigned to a Service
by OpenShift) and VRRP that is assigned by Neutron.

To address this OpenStackServiceNetwork setting is added to KuryrConfig
to serve as this broader OpenStack network CIDR. If not set
cluster-network-operator will set it to ServiceNetwork expanded by one
bit.