Update protobuf generated files to fix verify job
Seems like there was an upgrade of protoc (or libprotoc) in the CI's src
images. This causes the verify job to always fail as newly generated
files are different. This commit updates them to new version. Looks like
only the syntax of importing the packages is changed, so this commit
should not affect anything, beside unblocking the CI.