config/v1: add networkResourceGroupName for azure platform
The newtork resources like Virtual Network and Subnets are created currently for each cluster in the cluster's resource group.

To support reusing exisitng Virtual network and Subnets for OpenShift clusters, the cluster uses pre-existing ones in a separate resource group.

`networkResourceGroupName` is the resource group that contains the Virtual network and subnets for the cluster.

The resource group for network resources is the same as cluster unless specifically specified, therefore, this field is optional.

This is based on work in PR https://github.com/openshift/installer/pull/2441