operator/ingress: Add HTTP cookie capture
* operator/v1/types_ingress.go (CookieMatchType): New type.  Specify a type
of match (exact or prefix) for identifying cookies.
(IngressControllerCaptureHTTPCookie): New type.  Specify match type (using
the new CookieMatchType type), a cookie name or name prefix, and maximum
value length to use when capturing cookies for access logs.
(AccessLogging): Add HTTPCaptureCookies field, with type slice of
IngressControllerCaptureHTTPCookie.
* operator/v1/0000_50_ingress-operator_00-ingresscontroller.crd.yaml-merge-patch:
Patch the validation of the spec.logging.access.httpCaptureCookies field to
require the name subfield if the matchType subfield is "Exact" and require
the namePrefix subfield if matchType is "Prefix".
* operator/v1/0000_50_ingress-operator_00-ingresscontroller.crd.yaml:
* operator/v1/zz_generated.deepcopy.go:
* operator/v1/zz_generated.swagger_doc_generated.go: Regenerate.