create config for cloud-credential-operator
As one of the steps for implementing https://github.com/openshift/enhancements/blob/master/enhancements/installer/aws-permissions-check-bypass.md , create a config object to allow defining behavior for the cloud-credential-operator.

The CredentialsMode field can be used to inform CCO that it should avoid trying to determine the capabilities of the available cloud credentials (it is unreliable in at least AWS), and can also be used to place CCO in manual/disabled mode (where the user/admin is responsible for giving creds to the components requesting them).

The default value ("") for this field will put CCO into its current default of dynamically determining capabilities for the clouds where this is presently supported (AWS/Azure/GCP), and "passthrough" mode for all other platforms/clouds.