Bug 1852112: Make hostPrefix optional for the sake of non-(sdn/ovn) plugins
For plugins like which don't use the hostPrefix for anything, it does not
make sense to force customers to set specific values just to pass validation.
These plugins should be able to unset (0) the hostPrefix and not have to worry
about it. This patch makes hostPrefix optional. However note that if the
hostPrefix is set, the validation will happen and for the plugins like
OpenShiftSDN and OVNKubernetes which require hostPrefix, the validation
will happen irrespective of whether it is set or not.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>