Add `disabled` field to `HelmChartRepository`
The default Red Hat repo is delivered as a part of console operator payload:

https://github.com/openshift/console-operator/blob/master/manifests/01-helm.yaml

and since these resource are monitored by CVO, it is not possible to remove the default
configuration, and many customers are asking for exactly that.

In order to solve the customer request, `disabled` field is added to `HelmChartRepository`.
If set to true, that repo will be ignored/skipped in all relevant Helm operations.
At the same time, other cluster components (e.g. console UI) might use the new flag
to enable/disable some parts of UI.