Bug 1895053: Allow builds to mount the proxy CA
Add option that allows the cluster's proxy certificate authority to be
mounted via a buildah bind mount. When the CA is bind mounted, the
contents of `/etc/pki/ca-trust` are managed by the build container.
The folder and its subdirectories can be manipulated; however any changes
to the directory or its subdirectories are not persisted in the resulting
image.

For example, with this option set to true, a Dockerfile `RUN` instruction
that changes the ownership of `/etc/pki/ca-trust/extracted` will not have
the updated ownership appear in the build's output image.