operator/ingress: Add GCP Global Access for LBs
operator/v1/types_ingress.go:

Add a new type, `GCPLoadBalancerParameters`, with a single new field,
`ClientAccess`. This new `ClientAccess` field allows for users to
specify a `ClientAccess` setting of `Global` or `Local`, to accommodate
for the configuration of the GCP Internal Load Balancer Global Access
option.

This is in support of https://issues.redhat.com/browse/NE-517.

- operator/v1/0000_50_ingress-operator_00-ingresscontroller.crd.yaml:
- operator/v1/zz_generated.deepcopy.go:
- operator/v1/zz_generated.swagger_doc_generated.go:
  Regenerate.