operator/ingress: Add Ingress Controller Tunable Header Buffer Sizes
operator/v1/types_ingress.go:

Add a new `HTTPHeaderBuffer` field to `IngressControllerSpec`
that is of type `IngressControllerHTTPHeaderBuffer`.

`IngressControllerHTTPHeaderBuffer`, a new type, has the
following fields of type `int32`:

* `HeaderBufferBytes`
* `HeaderBufferMaxRewriteBytes`

This new type allows for users to configure the OpenShift router's
`tune.bufsize` and `tune.maxrewrite` values. See
https://cbonte.github.io/haproxy-dconv/2.0/configuration.html#tune.bufsize
&
https://cbonte.github.io/haproxy-dconv/2.0/configuration.html#tune.maxrewrite
for more information.

This is in support of https://issues.redhat.com/browse/NE-415.

- operator/v1/0000_50_ingress-operator_00-ingresscontroller.crd.yaml:
- operator/v1/zz_generated.deepcopy.go:
- operator/v1/zz_generated.swagger_doc_generated.go:
  Regenerate.