Include MTU option for KuryrConfig
When using a Provider Network for the Machines
the same MTU is attempted to be used for the
Tenant Networks of the Pods, which can be considered
too big by OpenStack Neutron. A new option is being included
to allow the user to set the desired MTU for the Pods
Network to avoid such scenario.