operator/dns: Describe default toleration
Update the godoc for the dnses.operator.openshift.io resource's
spec.nodePlacement.tolerations field to state that the DNS operator adds a
toleration for the "node-role.kubernetes.io/master" taint, that omitting
the toleration would be risky, and that the daemon controller adds some
additional tolerations.

This commit fixes bug 1951853.

https://bugzilla.redhat.com/show_bug.cgi?id=1951853

* operator/v1/types_dns.go (DNSNodePlacement): Fix description of the
Tolerations field.
* operator/v1/0000_70_dns-operator_00-custom-resource-definition.yaml:
* operator/v1/zz_generated.swagger_doc_generated.go: Regenerate.