config: add aws resourceTags to spec of infrastructure
Add the `spec.platformSpec.aws.resourceTags` field to the
infrastructure.config.openshift.io CRD. This is the user-facing
field for setting additional tags to be applied to AWS resources
created for the cluster. There will be a controller in
cluster-config-operator that syncs the field from the spec
to the status.