config/v1/types_cluster_operator: Clarify Available and Degraded severity
Available=False is really bad.  Possibly a page-at-midnight thing.
"Hey, your registry is down, so any new pods based on local images
will fail to launch" or "ingress is down, so your users cannot reach
you".  If it's not a page-at-midnight thing, it's at least going to be
the first batch of things admins should look at when they get the
from-the-spout alerts [1].

Degraded=True is not great, but you should be able to survive with
reduced quality-of-service until an admin wakes up in the morning.

[1]: https://docs.google.com/document/d/199PqyG3UsyXlwieHaqbGiWVa8eMWi8zzAn0YfcApr8Q/mobilebasic#h.rvrk9gcasjzh
     Rob Ewaschuk, My Philosophy on Alerting