config/v1/types_cluster_operator: Tighten up Upgradeable docs
"administrators should not upgrade" is way too strict.  This commit
relaxes that to point out that the message will tell you what's off
limits, and the CVO will enforce that (but you can force past the
CVO's guard).  The focus on minor updates is from [1].  Subsequently,
the CVO grew an internal check that even blocks patch bumps when they
are unlikely to succeed in the face of unsupported overrides [2].

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1797624
[2]: https://bugzilla.redhat.com/show_bug.cgi?id=1822844