config/v1/types_cluster_operator: Explain that conditions cover components
Some operators have no configured operands (e.g. the bare-metal
operator on non-metal platforms, or the image-registry operator when
the admins have configured managementState:Removed [1]).  Some
operators have many configured operands.  Operators writing
ClusterOperator conditions should not limit the conditions to speak
about just the operator, or just a particular operand.  Instead,
operators should speak about the component as a whole.  Is something
about the service that the component provides gone?  If so,
Available=False (midnight admin page).  Is something about the service
that the component provides not hitting its service-level objectives?
If so, Degraded=True (working-hours admin queue).  Doesn't matter if
that thing is "the operator is having trouble talking to the API to
figure out how the operands are doing" or "the operator is
super-happy, and sees that some operand is sad".  That's all stuff
that can be distinguished in the reason/message.

[1]: https://docs.openshift.com/container-platform/4.8/registry/configuring_registry_storage/configuring-registry-storage-baremetal.html#registry-removed_configuring-registry-storage-baremetal