Added forwarding policy for cluster DNS
Currently the upstream servers for the DNS forwarding are configured to
be queried randomly. This is the default behavior. The user currently
cannot specify a list of upstream servers to query in order. This can
adds a field where the user can specify a forwarding policy which can be
either random, round-robin or sequential. The policy defaults to random
to preserve the previous behavior.