Add Image Type to Azure Machine Provider Spec
With Azure marketplace images, some images require that the
subscription accept the terms of a purchase plan before using; and
that purchase plan information must be included in Azure API calls
when provisioning a VM. Images that do not require purchase plans
must not submit purchase plan info.

This adds an image type to distinguish between Marketplace images
with or without purchase plans. It could be used in the future to
support images from galleries.