Fix some discrepancies in field names in godoc
Fix some errors where the godoc for a struct field matches neither the
field identifier nor the json tag.

This commit fixes typos and other errors where it is clear that the godoc
was intended to refer to a field by name.  This commit does *not* update
godoc that was not obviously intended to refer to the field by name, *nor*
does this commit enforce uniformity in the use of either identifiers or
json tags.

* apiserver/v1/types_apirequestcount.go (PerUserAPIRequestCount): Change
"userName" to "username" in godoc.
* build/v1/types.go (BuildTriggerCause): Change "gitHubWebHook" to
"githubWebHook" in godoc.
(BuildConfigSpec): Add spaces between "//" comment markers and text.
(DockerStrategyOptions): Change "Args" to "buildArgs" in godoc.
(BuildLogOptions): Change "cointainer" to "container" and "noWait" to
"nowait" in godoc.
* cloudnetwork/v1/types.go (CloudPrivateIPConfigStatus): Change "condition"
to "conditions" in godoc.
* config/v1/types.go (GenericAPIServerConfig): Change "admissionConfig" to
"admission" in godoc.
* config/v1/types_apiserver.go (APIServerSpec): Change "servingCert" to
"servingCerts" in godoc.
* config/v1/types_cluster_version.go (ClusterCondition): Change "promQL" to
"promql" in godoc.
* config/v1/types_infrastructure.go (InfrastructureStatus): Change
"apiServerInternalURL" to "apiServerInternalURI" in godoc.
(GCPPlatformStatus): Change "resourceGroupName" to "projectID" in godoc.
* console/v1/types_console_link.go (ApplicationMenuSpec): Change "imageUrl"
to "imageURL" in godoc.
* image/v1/types.go (TagEventCondition): Change "LastTransitionTIme" to
"LastTransitionTime" in godoc.
* imageregistry/v1/types.go (ImageRegistryConfigStorageSwift): Change
"tenant" to "tenantID" in godoc.
* legacyconfig/v1/types.go (MasterConfig): Change "MasterVolumeConfig" to
"VolumeConfig" in godoc.
(MasterAuthConfig) Change "WebhookTokenAuthnConfig" to
"WebhookTokenAuthenticators" in godoc.
(RequestHeaderAuthenticationOptions): Change "GroupNameHeader" to
"GroupHeaders" in godoc.
(UserAgentMatchRule): Change "UserAgentRegex" to "Regex" in godoc.
(MasterNetworkConfig): Change "ServiceNetwork" to "ServiceNetworkCIDR" in
godoc.
(OAuthConfig): Add space between "//" comment marker and text.
(LDAPSyncConfig): Change "Host" to "URL" in godoc.
(RFC2307Config):
(AugmentedActiveDirectoryConfig): Change "GroupUIDAttributes" to
"GroupUIDAttribute" in godoc.
* machine/v1beta1/types_azureprovider.go (AzureMachineProviderSpec): Change
"InternalLoadBalancerName" to "InternalLoadBalancer name" in godoc.
* machine/v1beta1/types_gcpprovider.go (GCPEncryptionKeyReference): Change
"KMSKeyName" to "KMSKey" in godoc.
* machine/v1beta1/types_vsphereprovider.go (VSphereMachineProviderSpec):
Change "NumCPUs" to "NumCoresPerSocket" in godoc.
* networkoperator/v1/types_egressrouter.go (L4RedirectRule): Change "IP" to
"DestinationIP" in godoc.
* operator/v1/types_ingress.go (AWSLoadBalancerParameters): Change
"classicLoadBalancerParameters" to "classicLoadBalancer" and
"networkLoadBalancerParameters" to "networkLoadBalancer" in godoc.
* operator/v1/types_network.go (DefaultNetworkDefinition): Change
"openShiftSDNConfig" to "openshiftSDNConfig", "oVNKubernetesConfig" to
"ovnKubernetesConfig", and "KuryrConfig" to "kuryrConfig" in godoc.
(OVNKubernetesConfig): Change "geneve" to "genevePort" in godoc.
(NetFlowConfig): Change "netFlow" to "collectors" in godoc.
(SFlowConfig): Change "sFlowCollectors" to "collectors" in godoc.
(IPFIXConfig): Change "ipfixCollectors" to "collectors" in godoc.
(PolicyAuditConfig): Change "maxFilesSize" to "maxFileSize" in godoc.
* operator/v1/types_serviceca.go (ServiceCA): Add space between "//"
comment marker and text.
* operator/v1alpha1/types.go (NodeStatus): Change
"lastFailedDeploymentGenerationErrors" to "lastFailedDeploymentErrors" in
godoc.
* osin/v1/types.go (OAuthConfig): Add space between "//" comment marker and
text.
* route/v1/types.go (RouteIngress): Change "CanonicalHostname" to
"RouterCanonicalHostname" in godoc.
* sharedresource/v1alpha1/types_shared_configmap.go (SharedConfigMapSpec):
Add space between "//" comment marker and text.
* template/v1/types.go (BrokerTemplateInstanceSpec): Change
"templateinstance" to "templateInstance" and "bindingids" to "bindingIDs"
in godoc.
* apiserver/v1/apiserver.openshift.io_apirequestcount.yaml:
* apiserver/v1/zz_generated.swagger_doc_generated.go:
* build/v1/generated.proto:
* build/v1/zz_generated.swagger_doc_generated.go:
* cloudnetwork/v1/001-cloudprivateipconfig.crd.yaml:
* cloudnetwork/v1/generated.proto:
* cloudnetwork/v1/zz_generated.swagger_doc_generated.go:
* config/v1/0000_00_cluster-version-operator_01_clusterversion.crd.yaml:
* config/v1/0000_10_config-operator_01_apiserver.crd.yaml:
* config/v1/0000_10_config-operator_01_infrastructure.crd.yaml:
* config/v1/zz_generated.swagger_doc_generated.go:
* console/v1/0000_10_consolelink.crd.yaml:
* console/v1/zz_generated.swagger_doc_generated.go:
* image/v1/generated.proto:
* image/v1/zz_generated.swagger_doc_generated.go:
* imageregistry/v1/00-crd.yaml:
* imageregistry/v1/zz_generated.swagger_doc_generated.go:
* legacyconfig/v1/zz_generated.swagger_doc_generated.go:
* machine/v1beta1/zz_generated.swagger_doc_generated.go:
* networkoperator/v1/001-egressrouter.crd.yaml:
* networkoperator/v1/generated.proto:
* networkoperator/v1/zz_generated.swagger_doc_generated.go:
* operator/v1/0000_50_ingress-operator_00-ingresscontroller.crd.yaml:
* operator/v1/0000_70_cluster-network-operator_01.crd.yaml:
* operator/v1/zz_generated.swagger_doc_generated.go:
* operator/v1alpha1/zz_generated.swagger_doc_generated.go:
* route/v1/generated.proto:
* route/v1/zz_generated.swagger_doc_generated.go:
* template/v1/generated.proto:
* template/v1/zz_generated.swagger_doc_generated.go: Regenerate.