update AlibabaResourceReference to follow a union pattern
This change updates the AlibabaResourceReference struct to follow a
discriminated union pattern. It now contains a `type` field which will
indicate whether the resource reference is an ID or a list of tags. The
`ID` and `Tags` fields have been converted to pointers in following with
that pattern, and some constant values have been added for the various
types.