refactor ResourceGroup in AlibabaCloudMachineProviderConfig
This change converts the previous ResourceGroupID field into
ResourceGroup. This change is being made to convert that field from a
string into the AlibabaResourceReference type so that we can support
multiple types of lookup for the resource group. There are some cases,
notably during installation, where the ID may not be known, and in these
cases the Tags will be used to find the resource group. This change to
the structure gives us a method to detect this option when receiving the
API object in the machine api actuator.