update required fields for AlibabaCloudMachineProviderConfig
This change makes the `VpcID`, `SecurityGroups`, `VSwitch`, and
`ResourceGroups` fields required byu removing the `+optional` and
`omitempty` annotations. This change is being made to reflect the best
practices when deploying OpenShift on Alibaba Cloud.

When deploying on OpenShift, only the VPC mode is allowed. This means
that the user must specify a VPC ID or the instance creation will fail.
When creating instances in VPC mode, the user must specify a vSwitch or
the operation will fail.